requirements/base.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    make upgrade
#
analytics-python==1.2.9   # via -r requirements/base.in
attrs==20.3.0             # via edx-ace
certifi==2020.11.8        # via requests
cffi==1.14.3              # via cryptography
chardet==3.0.4            # via requests
coreapi==2.3.3            # via -r requirements/base.in, django-rest-swagger, openapi-codec
coreschema==0.0.4         # via coreapi
cryptography==3.2.1       # via pyjwt
defusedxml==0.6.0         # via python3-openid, social-auth-core
django-appconf==1.0.4     # via django-statici18n
django-compat==1.0.15     # via django-hijack
django-crum==0.7.9        # via edx-django-utils
django-debug-toolbar==3.1.1  # via -r requirements/base.in
django-extensions==3.0.9  # via -r requirements/base.in
django-filter==2.4.0      # via -r requirements/base.in
django-hijack==2.1.10     # via -r requirements/base.in
django-ratelimit==3.0.1   # via -r requirements/base.in
django-rest-swagger==2.2.0  # via -r requirements/base.in
django-sortedm2m==3.0.2   # via -r requirements/base.in
django-statici18n==2.0.1  # via -r requirements/base.in
django-storages==1.10.1   # via -r requirements/base.in
django-waffle==2.0.0      # via -r requirements/base.in, edx-django-utils, edx-drf-extensions
django-webpack-loader==0.7.0  # via -r requirements/base.in
django==2.2.24            # via -c requirements/constraints.txt, -r requirements/base.in, django-appconf, django-crum, django-debug-toolbar, django-filter, django-statici18n, django-storages, djangorestframework, drf-jwt, edx-ace, edx-auth-backends, edx-credentials-themes, edx-django-release-util, edx-django-sites-extensions, edx-django-utils, edx-drf-extensions, edx-i18n-tools, rest-condition, xss-utils
djangorestframework==3.12.2  # via -r requirements/base.in, django-rest-swagger, drf-jwt, edx-drf-extensions, rest-condition
drf-jwt==1.17.2           # via edx-drf-extensions
edx-ace==0.1.17           # via -r requirements/base.in
edx-auth-backends==3.1.0  # via -r requirements/base.in
edx-django-release-util==0.4.4  # via -r requirements/base.in
edx-django-sites-extensions==2.5.1  # via -r requirements/base.in
edx-django-utils==3.11.0  # via -r requirements/base.in, edx-drf-extensions, edx-rest-api-client
edx-drf-extensions==6.2.0  # via -r requirements/base.in
edx-i18n-tools==0.5.3     # via edx-credentials-themes
edx-opaque-keys==2.1.1    # via -r requirements/base.in, edx-drf-extensions
edx-rest-api-client==5.2.1  # via -r requirements/base.in
git+https://github.com/edx/credentials-themes.git@0.1.33#egg=edx_credentials_themes==0.1.33  # via -r requirements/base.in
future==0.18.2            # via pyjwkest
gitdb==4.0.5              # via gitpython
gitpython==3.1.11         # via transifex-client
idna==2.10                # via requests
itypes==1.2.0             # via coreapi
jinja2==2.11.2            # via coreschema
markdown==3.3.3           # via -r requirements/base.in
markupsafe==1.1.1         # via jinja2
mysqlclient==2.0.1        # via -r requirements/base.in
newrelic==5.22.1.152      # via -r requirements/base.in, edx-django-utils
oauthlib==3.1.0           # via requests-oauthlib, social-auth-core
openapi-codec==1.3.2      # via django-rest-swagger
path.py==12.5.0           # via edx-i18n-tools
path==15.0.0              # via path.py
pbr==5.5.1                # via stevedore
pillow==8.0.1             # via -r requirements/base.in
polib==1.1.0              # via edx-i18n-tools
psutil==5.7.3             # via edx-django-utils
pycparser==2.20           # via cffi
pycryptodomex==3.9.9      # via pyjwkest
pygments==2.7.2           # via -r requirements/base.in
pyjwkest==1.4.2           # via edx-drf-extensions
pyjwt[crypto]==1.7.1      # via drf-jwt, edx-auth-backends, edx-rest-api-client, social-auth-core
pymongo==3.11.0           # via edx-opaque-keys
python-dateutil==2.8.1    # via analytics-python, edx-ace, edx-drf-extensions
python-memcached==1.59    # via -r requirements/base.in
python-slugify==4.0.1     # via transifex-client
python3-openid==3.2.0     # via social-auth-core
pytz==2020.4              # via -r requirements/base.in, django
pyyaml==5.3.1             # via edx-django-release-util, edx-i18n-tools
requests-oauthlib==1.3.0  # via social-auth-core
requests==2.24.0          # via -c requirements/constraints.txt, -r requirements/base.in, analytics-python, coreapi, edx-drf-extensions, edx-rest-api-client, pyjwkest, requests-oauthlib, sailthru-client, slumber, social-auth-core, transifex-client
rest-condition==1.0.3     # via edx-drf-extensions
sailthru-client==2.2.3    # via edx-ace
semantic-version==2.8.5   # via edx-drf-extensions
simplejson==3.17.2        # via django-rest-swagger, sailthru-client
six==1.15.0               # via analytics-python, cryptography, django-compat, edx-ace, edx-auth-backends, edx-django-release-util, edx-drf-extensions, edx-i18n-tools, edx-opaque-keys, pyjwkest, python-dateutil, python-memcached, social-auth-app-django, social-auth-core, transifex-client
slumber==0.7.1            # via edx-rest-api-client
smmap==3.0.4              # via gitdb
git+https://github.com/python-social-auth/social-app-django.git@c00d23c2b45c3317bd35b15ad1b959338689cef8#egg=social-auth-app-django  # via -r requirements/base.in, edx-auth-backends
social-auth-core==3.2.0   # via -c requirements/constraints.txt, edx-auth-backends, social-auth-app-django
sqlparse==0.4.1           # via django, django-debug-toolbar
stevedore==3.2.2          # via edx-ace, edx-django-utils, edx-opaque-keys
text-unidecode==1.3       # via python-slugify
transifex-client==0.14.1  # via edx-credentials-themes
uritemplate==3.0.1        # via coreapi
urllib3==1.25.11          # via requests, transifex-client
xss-utils==0.1.3          # via -r requirements/base.in
jsonfield==3.0.0
importlib-metadata==1.6.0  # via markdown
zipp==1.2.0               # via importlib-metadata
